[MLIR][IR] Reduce overhead of Operation::getRegions
When returning the `MutableArrayRef` for the region list for
`getRegions()`, we take care to use a nullptr if there are no regions,
instead of using the region storage offset.  This extra branch has
slight performance overhead, and does not seem to increase safety
greatly as the array size will be 0.  Instead, use the region storage
address, even if it is 0 bytes in size, to construct the
`MutableArrayRef`. We already do this for for `getBlockOperands()`.